Разбивает код
