+++ public/index.html
@@ -13,26 +13,17 @@
 
     <header class="header">
       <h1 class="header__logo logo">Cinemaddict</h1>
-
-      <!-- Звание пользователя -->
     </header>
 
     <main class="main">
-      <!-- Меню -->
-
-      <!-- Сортировка -->
-
-      <!-- Контент -->
     </main>
 
 
     <footer class="footer">
       <section class="footer__logo logo logo--smaller">Cinemaddict</section>
-      <section class="footer__statistics">
-        <!-- Количество фильмов -->
-      </section>
+      <section class="footer__statistics"></section>
     </footer>
 
-    <script src="../src/main.js"></script>
+    <script src="./bundle.js"></script>
   </body>
 </html>
