+++ src/view/film-container.js
@@ -1,0 +1,10 @@
+const createFilmsContainerTemplate = () => {
+  return (
+    `<section class="films">
+    </section>`
+  );
+};
+
+export {
+  createFilmsContainerTemplate
+};
