+++ src/view/films-count.js
@@ -1,0 +1,9 @@
+const createFilmsCountTemplate = () => {
+  return (
+    `<p>130 291 movies inside</p>`
+  );
+};
+
+export {
+  createFilmsCountTemplate
+};
