+++ src/view/films-most-commented-list.js
@@ -1,0 +1,14 @@
+const createFilmsMostCommentedListTemplate = () => {
+  return (
+    `<section class="films-list--extra">
+      <h2 class="films-list__title">Most commented</h2>
+
+      <div class="films-list__container">
+      </div>
+    </section>`
+  );
+};
+
+export {
+  createFilmsMostCommentedListTemplate
+};
